ENH: rearrange DocumentNames in separate module
- remove after repo restructuring
