+++ src/event_model/generate/create_documents.py
@@ -227,7 +227,7 @@ from typing import Generic, Literal, TypeVar
 # NamedTuple cannot yet be generic.
 from typing_extensions import NamedTuple
 
-from event_model import DocumentNames
+from event_model.document_names import DocumentNames
 
 {imports}
